requirements.txt
matplotlib==3.6.3
numpy==1.24.2
pandas==1.5.3
Pillow==9.4.0
scikit_learn==1.2.1
seaborn==0.12.2
tensorflow==2.10.1
